Remove @behavior decorator, modify input type of @behaviors decorator
These decorators were previously applied when the module they were used
in was imported, unfortunately at that point the standard Polymer
behaviors may not have been imported yet so an element may have ended up
with some `undefined` behaviors which caused issues when attempting to
register the element with Polymer. To avoid this the remaining
@behaviors decorator now accepts a function that will be invoked to
obtain the array of behaviors which should be mixed into an element,
this function will be called when the element is registered with
Polymer.